bugfix: articles ID issue
Needed to create a system with a global varialbe for next ID, and to increment it at each new article created. No ID is used 2 times, every article will have a unic ID.

The issue was that everytime a new article was created, we checked the article table size and assigned + 1 to the ID, but if we delete items and than create new ones, with this logic, we will end up with ID conflicts.

For example we start with articles id 1, 2, and 3. We delete article with ID 2, we are left with ID 1 and 3. We create an article, its ID is set to table size + 1, so 2 + 1 = 3, same ID as the last item from the actual table. So now after creating the new article, the IDs of our articles are 1, 3, 3... conflict!